aidemy-bootstrap/planner/requirements.txt
pg8000==1.31.2
SQLAlchemy==2.0.36
cloud-sql-python-connector==1.16.0
google-genai
flask==3.1.0
jsonify==0.5
langchain_google_vertexai==2.0.13
langchain_core==0.3.34
pydantic==2.10.5
langgraph==0.2.70
google-cloud-pubsub==2.28.0
